Add mixin filters to the Diesel index implementation
- Introduced new mixin filters for improved query capabilities.
- Refactored existing code to utilize mixins for common query patterns.
