+++ devenv.nix
@@ -58,6 +58,9 @@
   languages.rust = {
     enable = true;
     channel = "nightly";
-    components = [ "rustc" "cargo" "clippy" "rustfmt" "rust-analyzer" "miri" ];
+    components = [
+      "rustc" "cargo" "clippy" "rustfmt" "rust-analyzer"
+      "miri"
+    ];
   };
 }
