+++ index/tests/clang-indexer.rs
@@ -80,7 +80,6 @@ fn mask_ref(reference: &Reference) -> Reference {
     Reference {
         from_decl: reference.from_decl,
         to_symbol: reference.to_symbol,
-        from_file: None,
         from_line: 1,
         from_col_start: 1,
         from_col_end: 1,
@@ -91,7 +90,6 @@ fn new_ref(from_decl: i32, to_symbol: i32) -> Reference {
     Reference {
         from_decl: DeclarationId::new(from_decl),
         to_symbol: SymbolId(to_symbol),
-        from_file: None,
         from_line: 1,
         from_col_start: 1,
         from_col_end: 1,
